feat(FI-1080) Rosetta core for network status endpoint

This MR proposes the following changes:
1. Move objects, identifiers, response and request types involved in the network/status endpoint to rosetta_core crate
2. Change Identifiers i64 to u64 where the standard proposes the type Int64 (>= 0) 

See merge request dfinity-lab/public/ic!16452

+++ rs/rosetta-api/icp_ledger/rosetta-integration-tests/src/rosetta_client.rs
@@ -32,7 +32,7 @@ impl RosettaClient {
         index: u64,
     ) -> Result<Option<Block>, reqwest::Error> {
         let block_identifier = PartialBlockIdentifier {
-            index: Some(index as i64),
+            index: Some(index),
             hash: None,
         };
         let request = BlockRequest {
